feat(db): 길이 분열 스윕 완결 + P2 키 규약 + 계보 수렴 — DB 표준화 로드맵 마감 (V2_18)
[P2 — 사용자 참조 키 규약 (권장안 승인)]
- docs/02-architecture/user-reference-key-policy.md 제정: 3계층 규약(소유참조=esntl_id+FK /
  행위자표기=loginId 스냅샷·FK금지 / 인증산출물=조회시점 키) — 전면 단일화 대신 계층 명문화
- rfsh_tk esntlId 단일 키잉 확정(전 경로 실측)+레거시 행 정리, User.changeUserId 제거(loginId 불변 선언)

[A2] fk_role_prgrm_map_* RENAME 보류 해제·원자 시행(라이브+V2_11+린터 화이트리스트) — 헌법 6조 비표준 FK명 완전 소멸
[A5] dev/local ddl-auto=validate 복원

[A1 — V2_18: 길이/타입 분열 스윕 (측정 워크플로우 3군 병렬, 전건 데이터 max 실측 무손실)]
- ALTER 56건: varchar(30) 클러스터·inst_cd 코드C7·authrt/role 참조그룹 20 원자정합(menu_crt_dtl
  12가 피참조 PK 실데이터 17자 수용불가던 잠재 파손 해소)·ymd→8 USING 정규화·_dt→timestamp·
  fax/telno/eml/pswd/link_url/ans_cn/sort_ordr 수렴 — 동일 컬럼명 cross-length 분열 그룹 0 달성
- uk_tb_onln_mnl_info 제약 승격+표준 개명, meta_standard_terms.eng_name DROP(정보량 0)
- 엔티티/DTO/서비스 ~40파일 동기화(2에이전트 팬아웃+메인 재검증) + 잠재버그 2건 해소
  (CnsltManage 29자 ISO 기록, MemoReport ISO 'T' 자가모순) + ymd 저장경로 normalizeYmd 봉쇄

[계보 수렴 — bootRun 실증]
- Flyway V2_6~V2_18 전량 history 등재(success 실측) + validate 통과 + api-docs 204 paths 추출
  → codegen:file/zod 재생성 diff 0(FE 계약 영향 없음 실증) → tsc exit 0
- 수렴이 잡은 결함 2건 즉시 수정: V2_14 시퀀싱 가드(V2_16 드랍 테이블 참조), V2_16/V2_18 멱등 가드
  (재실행 시 USING 리터럴·regex 가 변환된 타입과 충돌)

게이트: compileJava/compileTestJava·전 모듈 test(린터 포함)·frontend tsc 전부 그린
잔여(문서화): biz_cd 재모델링·etc_cd 스펙·TEXT 잔존 3·E2E 가비지 정리 — 태스크 기록 참조

+++ business-core/src/main/java/nuri/business/domain/deptjob/DeptJobBox.java
@@ -25,10 +25,10 @@ public class DeptJobBox extends BaseEntity {
     @Column(length = 20)
     private String deptId;
 
-    private Integer sortOrdr;
+    private Long sortOrdr;
 
     // 정적 팩토리 전용 생성자 (빌더 위임 대상)
-    private DeptJobBox(String deptTaskBoxId, String deptTaskBoxNm, String deptId, Integer sortOrdr) {
+    private DeptJobBox(String deptTaskBoxId, String deptTaskBoxNm, String deptId, Long sortOrdr) {
         this.deptTaskBoxId = deptTaskBoxId;
         this.deptTaskBoxNm = deptTaskBoxNm;
         this.deptId = deptId;
@@ -40,11 +40,11 @@ public class DeptJobBox extends BaseEntity {
      * 기존 {@code DeptJobBox.builder()...build()} 호출부와 호환.
      */
     @Builder
-    public static DeptJobBox create(String deptTaskBoxId, String deptTaskBoxNm, String deptId, Integer sortOrdr) {
+    public static DeptJobBox create(String deptTaskBoxId, String deptTaskBoxNm, String deptId, Long sortOrdr) {
         return new DeptJobBox(deptTaskBoxId, deptTaskBoxNm, deptId, sortOrdr);
     }
 
-    public void update(String deptTaskBoxNm, String deptId, Integer sortOrdr) {
+    public void update(String deptTaskBoxNm, String deptId, Long sortOrdr) {
         this.deptTaskBoxNm = deptTaskBoxNm;
         this.deptId = deptId;
         this.sortOrdr = sortOrdr;
