+++ business-core/src/main/java/nuri/business/service/code/dto/InstitutionCodeRecptnDto.java
@@ -119,7 +119,7 @@ public class InstitutionCodeRecptnDto {
     private String crtrYmd;
 
     @Schema(description = "정렬순서")
-    private Integer sortOrdr;
+    private Long sortOrdr;
 
     @Schema(description = "최초등록일시")
     private LocalDateTime crtDt;
